requirements/requirements_cpu.txt
absl-py @ file:///opt/conda/conda-bld/absl-py_1639803114343/work
aiofiles==22.1.0
aiohttp @ file:///tmp/build/80754af9/aiohttp_1646806365970/work
aiosignal @ file:///tmp/build/80754af9/aiosignal_1637843061372/work
albumentations==1.3.0
alembic==1.8.1
anyio==3.6.1
argcomplete==2.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
async-timeout @ file:///tmp/build/80754af9/async-timeout_1637851218186/work
attrs==22.1.0
autopage==0.5.1
backcall==0.2.0
backports.cached-property==1.0.2
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.4
boto3==1.24.71
botocore==1.27.71
bottle==0.12.23
brotlipy==0.7.0
cachetools @ file:///tmp/build/80754af9/cachetools_1619597386817/work
certifi @ file:///opt/conda/conda-bld/certifi_1663615672595/work/certifi
cffi @ file:///tmp/abs_98z5h56wf8/croots/recipe/cffi_1659598650955/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click @ file:///tmp/build/80754af9/click_1646056706450/work
cliff==4.0.0
cloudpickle @ file:///tmp/build/80754af9/cloudpickle_1632508026186/work
cmaes==0.8.2
cmd2==2.4.2
colorlog==6.7.0
cryptography @ file:///tmp/build/80754af9/cryptography_1652101588599/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
cytoolz==0.11.0
dacite==1.6.0
dask @ file:///tmp/abs_994957d9-ec12-411f-b953-c010f9d489d10hj3gz4k/croots/recipe/dask-core_1658513209934/work
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.5.1
distlib==0.3.6
dnspython==2.2.1
einops==0.5.0
empatches==0.2.0
entrypoints==0.4
eventlet==0.33.1
executing==1.0.0
fastjsonschema==2.16.1
fiftyone==0.16.6
fiftyone-brain==0.9.0
fiftyone-db==0.3.0
filelock==3.8.0
fonttools==4.25.0
frozenlist @ file:///home/builder/ci_310/frozenlist_1640790081390/work
fsspec @ file:///opt/conda/conda-bld/fsspec_1659972197723/work
future==0.18.2
glob2==0.7
google-auth @ file:///opt/conda/conda-bld/google-auth_1646735974934/work
google-auth-oauthlib @ file:///tmp/build/80754af9/google-auth-oauthlib_1617120569401/work
graphql-core==3.1.7
greenlet==1.1.3
grpcio @ file:///opt/conda/conda-bld/grpcio_1655801038495/work
gunicorn==20.1.0
h11==0.12.0
h2==4.1.0
hpack==4.0.0
httpcore==0.15.0
httpx==0.23.0
hypercorn==0.14.3
hyperframe==6.0.1
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
imagecodecs @ file:///home/builder/ci_310_2/imagecodecs_1642084528537/work
imageio @ file:///tmp/abs_cd920173-f360-47c5-97b0-bf4d1076d5d4dvic0oys/croots/recipe/imageio_1658785036907/work
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1648562408262/work
ipykernel==6.15.2
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.1.0
jsonschema==4.15.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter_client==7.3.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
kaleido==0.2.1
kiwisolver @ file:///opt/conda/conda-bld/kiwisolver_1653292039266/work
kornia==0.6.7
locket @ file:///opt/conda/conda-bld/locket_1652903118915/work
lxml==4.9.1
Mako==1.2.3
Markdown @ file:///home/builder/ci_310/markdown_1640791558615/work
MarkupSafe==2.1.1
matplotlib @ file:///opt/conda/conda-bld/matplotlib-suite_1660167928326/work
matplotlib-inline==0.1.6
mistune==2.0.4
mkl-fft==1.3.1
mkl-random @ file:///home/builder/ci_310/mkl_random_1641843545607/work
mkl-service==2.4.0
monai==1.0.0
mongoengine==0.20.0
motor==2.5.1
msgpack==1.0.4
multidict @ file:///opt/conda/conda-bld/multidict_1662369340274/work
munkres==1.1.4
nbclient==0.6.7
nbconvert==7.0.0
nbformat==5.4.0
ndjson==0.3.1
nest-asyncio==1.5.5
networkx @ file:///opt/conda/conda-bld/networkx_1657784097507/work
notebook==6.4.12
numpy @ file:///opt/conda/conda-bld/numpy_and_numpy_base_1653915516269/work
oauthlib @ file:///tmp/abs_08ngfezid4/croots/recipe/oauthlib_1659642459222/work
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
optuna==3.0.3
optuna-dashboard==0.7.3
optuna-fast-fanova==0.0.3
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas==1.4.4
pandocfilters==1.5.0
parso==0.8.3
partd @ file:///opt/conda/conda-bld/partd_1647245470509/work
patchify==0.2.3
patool==1.12
pbr==5.10.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
plotly==5.10.0
positional-encodings==6.0.1
pprintpp==0.4.0
prettytable==3.4.1
priority==2.0.0
prometheus-client==0.14.1
prompt-toolkit==3.0.31
protobuf==3.20.1
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1 @ file:///Users/ktietz/demo/mc3/conda-bld/pyasn1_1629708007385/work
pyasn1-modules==0.2.8
pycocotools==2.0.4
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments==2.13.0
PyJWT @ file:///opt/conda/conda-bld/pyjwt_1657544592787/work
pymongo==3.12.3
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing @ file:///opt/conda/conda-bld/pyparsing_1661452539315/work
pyperclip==1.8.2
pyrsistent==0.18.1
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-multipart==0.0.5
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets @ file:///tmp/build/80754af9/pywavelets_1648727875458/work
PyYAML==6.0
pyzmq==23.2.1
qtconsole==5.3.2
QtPy==2.2.0
qudida==0.0.4
ray==2.0.0
requests @ file:///opt/conda/conda-bld/requests_1657734628632/work
requests-oauthlib==1.3.0
retrying==1.3.3
rfc3986==1.5.0
rsa @ file:///tmp/build/80754af9/rsa_1614366226499/work
s3transfer==0.6.0
scikit-image @ file:///tmp/build/80754af9/scikit-image_1648214169132/work
scikit-learn==1.1.2
scipy @ file:///opt/conda/conda-bld/scipy_1661390393401/work
Send2Trash==1.8.0
sentinel==0.3.0
sip==4.19.13
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.41
sse-starlette==0.10.3
sseclient-py==1.7.2
stack-data==0.5.0
starlette==0.16.0
stevedore==4.0.1
strawberry-graphql==0.96.0
tabulate==0.8.10
tenacity==8.0.1
tensorboard @ file:///home/builder/stiwari/miniconda3/envs/tf_new_env/conda-bld/tensorboard_1661447826088/work/tensorboard-2.9.0-py3-none-any.whl
tensorboard-data-server @ file:///home/builder/ci_310/tensorboard-data-server_1640795016439/work/tensorboard_data_server-0.6.0-py3-none-manylinux2010_x86_64.whl
tensorboard-plugin-wit @ file:///home/builder/tkoch/workspace/tensorflow/tensorboard-plugin-wit_1658918494740/work/tensorboard_plugin_wit-1.8.1-py3-none-any.whl
tensorboardX==2.5.1
terminado==0.15.0
threadpoolctl==3.1.0
tifffile @ file:///tmp/build/80754af9/tifffile_1627275862826/work
tinycss2==1.1.1
toml==0.10.2
toolz @ file:///tmp/build/80754af9/toolz_1636545406491/work
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
tornado @ file:///opt/conda/conda-bld/tornado_1662061693373/work
tqdm==4.64.1
traitlets==5.3.0
typing_extensions @ file:///tmp/abs_ben9emwtky/croots/recipe/typing_extensions_1659638822008/work
tzdata==2022.2
tzlocal==4.2
universal-analytics-python3==1.1.1
urllib3 @ file:///tmp/abs_5dhwnz6atv/croots/recipe/urllib3_1659110457909/work
virtualenv==20.16.5
voxel51-eta==0.8.1
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug @ file:///opt/conda/conda-bld/werkzeug_1645628268370/work
widgetsnbextension==4.0.3
wrapt==1.14.1
wsproto==1.2.0
xmltodict==0.13.0
yarl @ file:///opt/conda/conda-bld/yarl_1661437085904/work
zipp @ file:///opt/conda/conda-bld/zipp_1652341764480/work
